# [SPARK-32914][SQL] Avoid constructing dataType multiple times