--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/collectionOperations.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/collectionOperations.scala
@@ -3502,7 +3502,6 @@ case class ArrayIntersect(left: Expression, right: Expression) extends ArrayBina
   with ComplexTypeMergingExpression {
 
   private lazy val internalDataType: DataType = {
-    dataTypeCheck
     ArrayType(elementType,
       left.dataType.asInstanceOf[ArrayType].containsNull &&
         right.dataType.asInstanceOf[ArrayType].containsNull)
@@ -3746,12 +3745,7 @@ case class ArrayIntersect(left: Expression, right: Expression) extends ArrayBina
 case class ArrayExcept(left: Expression, right: Expression) extends ArrayBinaryLike
   with ComplexTypeMergingExpression {
 
-  private lazy val internalDataType: DataType = {
-    dataTypeCheck
-    left.dataType
-  }
-
-  override def dataType: DataType = internalDataType
+  override def dataType: DataType = left.dataType
 
   @transient lazy val evalExcept: (ArrayData, ArrayData) => ArrayData = {
     if (TypeUtils.typeWithProperEquals(elementType)) {