@cloud-fan 2020-09-18
do we really need such an invasive change? If there is a specific expression that calls `dataType` many times, let's fix that expression only. Or if this can bring significant end-to-end perf speedup, we can consider accepting it.