@cloud-fan 2020-09-18
Or if an expression has a very complicated `def dataType`, can we change it to `lazy val dataType`?